requirements.txt
alpha-vantage==2.1.1
appnope==0.1.0
asn1crypto==1.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
dash==1.3.1
dash-core-components==1.2.1
dash-html-components==1.0.1
dash-renderer==1.1.0
dash-table==4.3.0
dateparser==0.7.2
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.17.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pipreqs==0.4.9
plotly==4.1.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycoingecko==0.4.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyrsistent==0.15.4
python-binance==0.7.4
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.5
regex==2019.8.19
requests==2.22.0
retrying==1.3.3
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
Twisted==19.7.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
yarg==0.1.9
zipp==0.6.0
zope.interface==4.6.0
